requirements.txt
pyannote.audio
openai-whisper
torchaudio
torch
python-Levenshtein
pydu
speechbrain
silero-vad
seaborn
psycopg2
spacy
transformers
statsmodels
ipython
datasets
python-dotenv

boto3